+++ src/types/model/table.ts
@@ -1,0 +1,13 @@
+import type { CustomTypeModelFieldType } from "./types"
+
+/**
+ * A table custom type field.
+ *
+ * More details: {@link https://prismic.io/docs/table}
+ */
+export interface CustomTypeModelTableField {
+	type: typeof CustomTypeModelFieldType.Table
+	config?: {
+		label?: string | null
+	}
+}