+++ test/types/customType-table.types.ts
@@ -1,0 +1,38 @@
+import { expectNever, expectType } from "ts-expect"
+
+import type * as prismicTICustomTypes from "@prismicio/types-internal/lib/customtypes"
+
+import * as prismic from "../../src"
+
+;(value: prismic.CustomTypeModelTableField): true => {
+	switch (typeof value) {
+		case "object": {
+			if (value === null) {
+				expectNever(value)
+			}
+
+			return true
+		}
+
+		default: {
+			return expectNever(value)
+		}
+	}
+}
+
+expectType<prismic.CustomTypeModelTableField>({
+	type: prismic.CustomTypeModelFieldType.Table,
+	config: {
+		label: "string",
+	},
+})
+
+/**
+ * `@prismicio/types` extends `@prismicio/types-internal`
+ */
+expectType<prismic.CustomTypeModelTableField>({} as prismicTICustomTypes.Table)
+
+/**
+ * `@prismicio/types-internal` extends `@prismicio/types`
+ */
+expectType<prismicTICustomTypes.Table>({} as prismic.CustomTypeModelTableField)